LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-across-slot-expected.txt


Open popover This button is where focus should land after traversing this popover

FAIL Popover focus behaviors in form control elements assert_equals: Focus should move forward from invoker to target1 expected Element node <input type="checkbox" id="target1"></input> but got Element node <body><div popover="" id="popover">
  <div id="host">
   ...
